pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fuzion"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["rich", "playwright", "pyyaml"]

[project.scripts]
fuzion = "fuzion.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
